move exeCommand into subdir
